requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2
pdfpages==0.1.0
argparse==1.4.0
plotly==5.22.0
